make edit pipettes work and refactor modal files
