+++ protocol-designer/src/components/ProtocolEditor.js
@@ -5,7 +5,7 @@ import ConnectedNav from '../containers/ConnectedNav'
 import ConnectedSidebar from '../containers/ConnectedSidebar'
 import ConnectedTitleBar from '../containers/ConnectedTitleBar'
 import ConnectedMainPanel from '../containers/ConnectedMainPanel'
-import ConnectedNewFileModal from '../containers/ConnectedNewFileModal'
+import NewFileModal from './modals/NewFileModal'
 import FileUploadErrorModal from './modals/FileUploadErrorModal'
 import AnalyticsModal from './modals/AnalyticsModal'
 import {PortalRoot as MainPageModalPortalRoot} from '../components/portals/MainPageModalPortal'
@@ -30,7 +30,7 @@ export default function ProtocolEditor () {
 
           <div className={styles.main_page_content}>
             <AnalyticsModal />
-            <ConnectedNewFileModal useProtocolFields />
+            <NewFileModal useProtocolFields />
             <FileUploadErrorModal />
             {/* TODO: Ian 2018-06-28 All main page modals will go here */}
             <MainPageModalPortalRoot />
