+++ protocol-designer/src/components/modals/EditPipettesModal/index.js
@@ -1,6 +1,8 @@
 // @flow
 import type {ElementProps} from 'react'
 import {connect} from 'react-redux'
+import isEmpty from 'lodash/isEmpty'
+import last from 'lodash/last'
 import mapValues from 'lodash/mapValues'
 
 import {uuid} from '../../../utils'
@@ -10,14 +12,16 @@ import {
   actions as stepFormActions,
   selectors as stepFormSelectors,
 } from '../../../step-forms'
-import NewFileModal from '../NewFileModal'
+import FilePipettesModal from '../FilePipettesModal'
 import type {BaseState, ThunkDispatch} from '../../../types'
-import type {PipetteEntities, PipetteOnDeck} from '../../../step-forms'
+import type {PipetteOnDeck} from '../../../step-forms'
+import type {StepIdType} from '../../../form-types'
 
-type Props = ElementProps<typeof NewFileModal>
+type Props = ElementProps<typeof FilePipettesModal>
 
 type SP = {
-  _prevPipettes: PipetteEntities,
+  _prevPipettes: {[pipetteId: string]: PipetteOnDeck},
+  _orderedSteps: Array<StepIdType>,
 }
 
 type OP = {
@@ -28,16 +32,20 @@ const mapSTP = (state: BaseState): SP => {
   const initialPipettes = stepFormSelectors.getPipettesForEditPipetteForm(state)
   return {
     initialPipetteValues: initialPipettes,
-    _prevPipettes: stepFormSelectors.getPipetteInvariantProperties(state),
+    _prevPipettes: stepFormSelectors.getInitialDeckSetup(state).pipettes, // TODO: Ian 2019-01-02 when multi-step editing is supported, don't use initial deck state. Instead, show the pipettes available for the selected step range
+    _orderedSteps: stepFormSelectors.getOrderedSteps(state),
   }
 }
 
-const makeUpdatePipettes = (prevPipettes, dispatch, closeModal) =>
-  ({pipettes}) => {
+const makeUpdatePipettes = (prevPipettes, orderedSteps, dispatch, closeModal) =>
+  ({pipettes: newPipetteArray}) => {
     const prevPipetteIds = Object.keys(prevPipettes)
-    let usedPrevPipettes = [] // IDs of pipettes in prevPipettes that were already used
+    let usedPrevPipettes: Array<string> = [] // IDs of pipettes in prevPipettes that were already put into nextPipettes
     let nextPipettes: {[pipetteId: string]: PipetteOnDeck} = {}
-    pipettes.forEach(newPipette => {
+
+    // from array of pipettes from Edit Pipette form (with no IDs),
+    // assign IDs and populate nextPipettes
+    newPipetteArray.forEach(newPipette => {
       if (newPipette && newPipette.name && newPipette.tiprackModel) {
         const candidatePipetteIds = prevPipetteIds.filter(id => {
           const prevPipette = prevPipettes[id]
@@ -63,12 +71,31 @@ const makeUpdatePipettes = (prevPipettes, dispatch, closeModal) =>
     dispatch(steplistActions.changeSavedStepForm({
       stepId: INITIAL_DECK_SETUP_STEP_ID,
       update: {
-        pipetteLocationUpdate: mapValues(nextPipettes, p => p.mount),
+        pipetteLocationUpdate: mapValues(nextPipettes, (p: PipetteOnDeck) => p.mount),
       },
     }))
 
+    const pipetteIdsToDelete: Array<string> = Object.keys(prevPipettes).filter(id => !(id in nextPipettes))
+    const substitutionMap = pipetteIdsToDelete.reduce((acc: {[string]: string}, deletedId: string): {[string]: string} => {
+      const deletedPipette = prevPipettes[deletedId]
+      const replacementId = Object.keys(nextPipettes)
+        .find(newId => nextPipettes[newId].mount === deletedPipette.mount)
+      return (replacementId && replacementId !== -1)
+        ? {...acc, [deletedId]: replacementId}
+        : acc
+    }, {})
+
+    // substitute deleted pipettes with new pipettes on the same mount, if any
+    if (!isEmpty(substitutionMap) && orderedSteps.length > 0) {
+      // NOTE: using start/end here is meant to future-proof this action for multi-step editing
+      dispatch(stepFormActions.substituteStepFormPipettes({
+        substitutionMap,
+        startStepId: orderedSteps[0],
+        endStepId: last(orderedSteps),
+      }))
+    }
+
     // delete any pipettes no longer in use
-    const pipetteIdsToDelete = Object.keys(prevPipettes).filter(id => !(id in nextPipettes))
     if (pipetteIdsToDelete.length > 0) {
       dispatch(stepFormActions.deletePipettes(pipetteIdsToDelete))
     }
@@ -78,8 +105,8 @@ const makeUpdatePipettes = (prevPipettes, dispatch, closeModal) =>
 
 const mergeProps = (stateProps: SP, dispatchProps: {dispatch: ThunkDispatch<*>}, ownProps: OP): Props => {
   const {dispatch} = dispatchProps
-  const {_prevPipettes, ...passThruStateProps} = stateProps
-  const updatePipettes = makeUpdatePipettes(_prevPipettes, dispatch, ownProps.closeModal)
+  const {_prevPipettes, _orderedSteps, ...passThruStateProps} = stateProps
+  const updatePipettes = makeUpdatePipettes(_prevPipettes, _orderedSteps, dispatch, ownProps.closeModal)
 
   return {
     ...ownProps,
@@ -90,4 +117,4 @@ const mergeProps = (stateProps: SP, dispatchProps: {dispatch: ThunkDispatch<*>},
   }
 }
 
-export default connect(mapSTP, null, mergeProps)(NewFileModal)
+export default connect(mapSTP, null, mergeProps)(FilePipettesModal)
