+++ protocol-designer/src/step-forms/actions/index.js
@@ -1,6 +1,7 @@
 // @flow
 import mapValues from 'lodash/mapValues'
 import pick from 'lodash/pick'
+import type {StepIdType} from '../../form-types'
 
 export type CreatePipettesAction = {
   type: 'CREATE_PIPETTES',
@@ -31,14 +32,21 @@ export const deletePipettes = (payload: $PropertyType<DeletePipettesAction, 'pay
   payload,
 })
 
-export type ModifyPipettesTiprackAssignmentAction = {
-  type: 'MODIFY_PIPETTES_TIPRACK_ASSIGNMENT',
+export type SubstituteStepFormPipettesAction = {
+  type: 'SUBSTITUTE_STEP_FORM_PIPETTES',
   payload: {
-    [pipetteId: string]: string, // new assigned tiprack model
+    // step range to modify (inclusive)
+    startStepId: StepIdType,
+    endStepId: StepIdType,
+
+    // old pipette id -> new id
+    substitutionMap: {[oldPipetteId: string]: string},
   },
 }
 
-export const modifyPipettesTiprackAssignment = (payload: $PropertyType<ModifyPipettesTiprackAssignmentAction, 'payload'>): ModifyPipettesTiprackAssignmentAction => ({
-  type: 'MODIFY_PIPETTES_TIPRACK_ASSIGNMENT',
+export const substituteStepFormPipettes = (
+  payload: $PropertyType<SubstituteStepFormPipettesAction, 'payload'>
+): SubstituteStepFormPipettesAction => ({
+  type: 'SUBSTITUTE_STEP_FORM_PIPETTES',
   payload,
 })
